Check if Celeste.exe_ exists before moving it back to Celeste.exe
In the case where Celeste is already running and trying to run Celeste.exe --vanilla, an error saying 'cannot access the file because it is used by another process' appears when trying to move Celeste.exe to Celeste.exe_. Then, the finally block is called and tries to move the file back (except it wasn't actually moved in the first place). The error that ends up being logged is 'Celeste.exe_ not found', which is quite confusing.

+++ Celeste.Mod.mm/Mod/Everest/BOOT.cs
@@ -138,7 +138,8 @@ namespace Celeste.Mod {
                         ad.DoCallBack(cal.GetMethod("Run").CreateDelegate(typeof(CrossAppDomainDelegate)) as CrossAppDomainDelegate);
 
                     } finally {
-                        File.Move(everestPath + "_", everestPath);
+                        if (File.Exists(everestPath + "_"))
+                            File.Move(everestPath + "_", everestPath);
                     }
 
                     // Luckily the newly loaded vanilla Celeste.exe becomes the executing assembly from now on.
